use `bfloat16` instead of `float16` in forward pas

+++ src/pyjuice/layer/sum_layer.py
@@ -494,8 +494,8 @@ class SumLayer(Layer, nn.Module):
 
                 if use_fp16 == 1:
                     # Built-in matmul kernel of triton + float16
-                    epars_fp16 = epars.to(tl.float16)
-                    emars_fp16 = emars_sub.to(tl.float16)
+                    epars_fp16 = epars.to(tl.bfloat16)
+                    emars_fp16 = emars_sub.to(tl.bfloat16)
                     nmars = tl.dot(epars_fp16, emars_fp16).to(tl.float32)
                 else:
                     # Built-in matmul kernel of triton + float32
